fixed rubout of macro invocations: goto tables and q-registers are allocated on the C++ call stack and configured to not emit undo tokens
this introduces additional logic but has the huge advantage that the tables can be freed after the macro invocation.
if undo tokens were emitted, the tables had to be kept in the undo stack so they can be restored during rubout.
this however would be both complicated and unnecessarily inefficient since the tables would reach their initial state
during rubout and be deallocated anyways.
 * similar (but not strictly necessary optimizations) can be performed for macro invocations
 * also wrapper Q-Register setting/getting -> will allow a custom "*" register getter (e.g. calculates buffer position
   on the fly)