throw error when macro returns with an unterminated command
it was not possible to inherit the parser state of the last command in
a macro. However, it was possible to exit the macro when its
last command was not properly terminated.
There is no useful conscious application of this behaviour.
 * If something like an uncomplete E-Command was last in the
   macro, there was absolutely no effect. However this points
   to a macro programming error.
 * If the last command started a string parameter but did not
   terminate it, the command might be (partially) executed.
   However the State's done() method wasn't called, which means
   that some commands will not execute at all. Again there's
   no useful application of this.
When on the other hand, the last command was not terminated
by accident, this was not reported to the user.