define scintilla.a as a phony target
when hacking Scintilla, it is useful to recompile it as
necessary. Since SciTECO calls the Scintilla/Scinterm Makefile
recursively, we do not know the libraries dependencies in
SciTECO's build system. It therefore makes sense to define
externally built targets as phony, so the recursive make
is called every time scintilla.a is required. If scintilla.a
is already up to date, the additional recursive make call
won't hurt.