initialize some variables in teco_state_expectqreg_input() to silence the Clang static analyzer
There was the possibility that an unitialized value is passed into a got_register_cb().
This could only happen in normal parse mode and it's unlikely to have caused problems in practice
as all callbacks should not rely on qreg being valid in parse-only mode.
This sort of bug would have also been catched by Valgrind.

The rest of the scan-build-reported warnings are bogus
(caused by not being g_auto-aware).