UNIX curses: some clarifications on what is done during teco_interface_init_screen()
User messages printed in interative mode aren't currently fully preserved on stdout/stderr
since they are redirected to /dev/null.
Only messages that are not flushed out will be preserved.
Unless you redirect stdout/stderr of SciTECO of course since in this case no redirection
is necessary.

This is probably tolerable esp. once we support multi-line messages in the UIs.
At least it would be tricky to work around.